requirements.txt
flask
python-dotenv
langchain
langchain-groq